+++ test/json/testApiAddNodeJSON.cpp
@@ -77,7 +77,7 @@ TEST_F(ApiAddNodeJSON, AddNestedObjectWithArrayToObject)
     auto api = prepareApiWithJson("test_5.json");
     std::map<std::string, NodeApi> newObject;
     newObject.emplace("b", true);
-    newObject.emplace("a", std::vector<NodeApi>{ { 0 }, { 100 }, { 200 } });
+    newObject.emplace("a", std::vector<NodeApi>{{ 0 }, { 100 }, { 200 }});
 
     const auto begin = std::chrono::high_resolution_clock::now();
     bool result = api->addNodeIntoObject({ "person" }, "newValues", { newObject });
@@ -94,7 +94,7 @@ TEST_F(ApiAddNodeJSON, AddNestedObjectWithArrayToObject)
 TEST_F(ApiAddNodeJSON, AddArrayToObject)
 {
     auto api = prepareApiWithJson("test_4.json");
-    std::vector<NodeApi> newArray{ { 232 }, { 234234 }, { 0 }, { 100 } };
+    std::vector<NodeApi> newArray{{ 232 }, { 234234 }, { 0 }, { 100 }};
 
     const auto begin = high_resolution_clock::now();
     bool result = api->addNodeIntoObject({ "person2", "address" }, "dataArrayNode", { newArray });
@@ -112,10 +112,10 @@ TEST_F(ApiAddNodeJSON, AddSimpleNodeIntoArray)
 {
     auto api = prepareApiWithJson("test_7.json");
 
-    bool result = api->addNodeIntoArray({ "employees", size_t(0), "data", size_t(0) }, NodeApi{ .value = 4 });
+    bool result = api->addNodeIntoArray({ "employees", uint(0), "data", uint(0) }, NodeApi{ .value = 4 });
     ASSERT_TRUE(result);
 
-    result = api->insertNodeIntoArray({ "employees", size_t(1), "data", size_t(0) }, 1, NodeApi{ .value = "c c" });
+    result = api->insertNodeIntoArray({ "employees", uint(1), "data", uint(0) }, 1, NodeApi{ .value = "c c" });
     ASSERT_TRUE(result);
 
     std::string json = api->objectToJsonString().value();
@@ -131,7 +131,7 @@ TEST_F(ApiAddNodeJSON, AddObjectIntoArray)
     newObjectNode.emplace("aa", "bb");
     newObjectNode.emplace("cc", 12);
 
-    bool result = api->insertNodeIntoArray({ "employees" , size_t(1), "data", size_t(0) }, 1, { newObjectNode });
+    bool result = api->insertNodeIntoArray({ "employees" , uint(1), "data", uint(0) }, 1, { newObjectNode });
     ASSERT_TRUE(result);
 
     std::string json = api->objectToJsonString().value();
@@ -143,12 +143,12 @@ TEST_F(ApiAddNodeJSON, AddObjectIntoArray)
 TEST_F(ApiAddNodeJSON, AddArrayIntoArray)
 {
     auto api = prepareApiWithJson("test_7.json");
-    std::vector<NodeApi> arr1{ { 1 }, { 2 }, { 3 } };
-    std::vector<NodeApi> arr2{ { "aa" }, { "b" } };
-    std::vector<NodeApi> arr3{ { true }, { false } };
-    std::vector<NodeApi> newArrayNode{ { arr1 }, { arr2 }, { arr3 } };
+    std::vector<NodeApi> arr1{{ 1 }, { 2 }, { 3 }};
+    std::vector<NodeApi> arr2{{ "aa" }, { "b" }};
+    std::vector<NodeApi> arr3{{ true }, { false }};
+    std::vector<NodeApi> newArrayNode{{ arr1 }, { arr2 }, { arr3 }};
 
-    bool result = api->addNodeIntoArray({ "employees", size_t(1), "data" }, { newArrayNode });
+    bool result = api->addNodeIntoArray({ "employees", uint(1), "data" }, { newArrayNode });
     ASSERT_TRUE(result);
 
     std::string json = api->objectToJsonString().value();
